feat: implemented interactions for ActionSet actions
